Move views to feature import export unit test
